requirements.txt
Babel==2.8.0
Click==7.0
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Cors==3.0.8
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
python-dotenv==0.12.0
pytz==2019.3
repoze.lru==0.7
Routes==2.4.1
six==1.14.0
Werkzeug==1.0.0
WTForms==2.2.1
xmlrunner==1.7.7
autoenv==1.0.0
